+++ src/main/scala/starter/starter.scala
@@ -1,30 +1,38 @@
 package scalajs.starter
 
 import scala.scalajs.js
-import js.annotation.{ JSExport, JSExportTopLevel }
+import js.annotation.{ JSExportAll, JSGlobalScope }
 
-@JSExportTopLevel("Starter")
-object Starter {
-  @JSExport
+@js.native
+@JSGlobalScope
+object Globals extends js.Object {
+  var globalThis: js.Dynamic = js.native
+  // var StarterClassInstance: StarterClass = js.native
+  var Starter: StarterObject = js.native
+}
+
+sealed trait StarterObject
+
+@JSExportAll
+object Starter extends StarterObject {
   val someval: String = "some starter object value"
   
-  @JSExport
   def greet(name: String): Unit = {
-    println(s"Hello, $name!")
+    Globals.globalThis.console.log(s"Hello, $name!")
   }
 
   def main(args: Array[String]): Unit = {
-    println("Hello from Scala.js starter object!")
+    Globals.globalThis.console.log("Hello from Scala.js starter object!")
+    Globals.Starter = this;
+    // Globals.StarterClassInstance = new StarterClass();
   }
 }
 
-@JSExportTopLevel("StarterClass")
+@JSExportAll
 class StarterClass {
-  @JSExport
   val someval: String = "some starter class value"
   
-  @JSExport
   def greet(name: String): Unit = {
-    println(s"Hello, $name!")
+    Globals.globalThis.console.log(s"Hello, $name!")
   }
 }